// client/src/module_bindings/types.ts
// THIS FILE IS AUTOMATICALLY GENERATED BY SPACETIMEDB. EDITS TO THIS FILE
// WILL NOT BE SAVED. MODIFY TABLES IN YOUR MODULE SOURCE CODE INSTEAD.

/* eslint-disable */
/* tslint:disable */
import {
  TypeBuilder as __TypeBuilder,
  t as __t,
  type AlgebraicTypeType as __AlgebraicTypeType,
  type Infer as __Infer,
} from "spacetimedb";

export const Message = __t.object("Message", {
  id: __t.u64(),
  sender: __t.identity(),
  text: __t.string(),
  sentAt: __t.timestamp(),
});
export type Message = __Infer<typeof Message>;

export const User = __t.object("User", {
  identity: __t.identity(),
  name: __t.string(),
});
export type User = __Infer<typeof User>;